+++ app/src/main/AndroidManifest.xml
@@ -22,20 +22,22 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme"
         tools:ignore="GoogleAppIndexingWarning">
-        <activity android:name=".Lighting"></activity>
-        <activity android:name=".Chandelier" />
-        <activity android:name=".E_Device" />
-        <activity android:name=".Socket" />
-        <activity android:name=".Generator" />
-        <activity android:name=".Security" />
-        <activity android:name=".Shutter_Door" />
-        <activity android:name=".Cooler" />
-        <activity android:name=".E_panel" />
-        <activity android:name=".Antenna" />
-        <activity android:name=".Wiring" />
-        <activity android:name=".Santeral" />
-        <activity android:name=".Iphon" />
-        <activity android:name=".Camera" />
+        <activity android:name=".Activity.Contact"></activity>
+        <activity android:name=".Activity.Map" />
+        <activity android:name=".Services.Lighting" />
+        <activity android:name=".Services.Chandelier" />
+        <activity android:name=".Services.E_Device" />
+        <activity android:name=".Services.Socket" />
+        <activity android:name=".Services.Generator" />
+        <activity android:name=".Services.Security" />
+        <activity android:name=".Services.Shutter_Door" />
+        <activity android:name=".Services.Cooler" />
+        <activity android:name=".Services.E_panel" />
+        <activity android:name=".Services.Antenna" />
+        <activity android:name=".Services.Wiring" />
+        <activity android:name=".Services.Santeral" />
+        <activity android:name=".Services.Iphon" />
+        <activity android:name=".Services.Camera" />
         <activity android:name=".Asansor" />
         <activity
             android:name=".Panels.Sms"
