+++ app/src/main/java/ir/mseif/app/com/fzm/Activity/Map.java
@@ -1,4 +1,4 @@
-package ir.mseif.app.com.fzm;
+package ir.mseif.app.com.fzm.Activity;
 
 import android.content.Context;
 import android.content.Intent;
@@ -6,24 +6,22 @@ import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.view.View;
 
+import ir.mseif.app.com.fzm.R;
 import uk.co.chrisjenx.calligraphy.CalligraphyContextWrapper;
 
-public class Wiring extends AppCompatActivity {
+public class Map extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_wiring);
+        setContentView(R.layout.activity_map);
     }
 
 
 
 
-    // Go To Time Activity
-    public void Go_To_Time (View v){
-        Intent intent = new Intent(getApplicationContext(), MainActivity.class);
-        startActivity(intent);
-    }
+
+
 
 
     @Override
