+++ app/src/main/java/ir/mseif/app/com/fzm/Services/Generator.java
@@ -1,4 +1,4 @@
-package ir.mseif.app.com.fzm;
+package ir.mseif.app.com.fzm.Services;
 
 import android.content.Context;
 import android.content.Intent;
@@ -6,17 +6,26 @@ import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.view.View;
 
+import ir.mseif.app.com.fzm.MainActivity;
+import ir.mseif.app.com.fzm.Activity.Map;
+import ir.mseif.app.com.fzm.R;
 import uk.co.chrisjenx.calligraphy.CalligraphyContextWrapper;
 
-public class Iphon extends AppCompatActivity {
+public class Generator extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_iphon);
+        setContentView(R.layout.activity_generator);
     }
 
 
+
+    // Go To Time Activity
+    public void Go_To_Location (View v){
+        Intent intent = new Intent(getApplicationContext(), Map.class);
+        startActivity(intent);
+    }
 
 
     // Go To Time Activity
