+++ app/src/main/java/ir/mseif/app/com/fzm/Services/Lighting.java
@@ -1,4 +1,4 @@
-package ir.mseif.app.com.fzm;
+package ir.mseif.app.com.fzm.Services;
 
 import android.content.Context;
 import android.content.Intent;
@@ -6,6 +6,9 @@ import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.view.View;
 
+import ir.mseif.app.com.fzm.MainActivity;
+import ir.mseif.app.com.fzm.Activity.Map;
+import ir.mseif.app.com.fzm.R;
 import uk.co.chrisjenx.calligraphy.CalligraphyContextWrapper;
 
 public class Lighting extends AppCompatActivity {
@@ -17,6 +20,12 @@ public class Lighting extends AppCompatActivity {
     }
 
 
+
+    // Go To Time Activity
+    public void Go_To_Location (View v){
+        Intent intent = new Intent(getApplicationContext(), Map.class);
+        startActivity(intent);
+    }
 
 
     // Go To Time Activity
