+++ app/src/main/java/ir/mseif/app/com/fzm/Services/Shutter_Door.java
@@ -1,4 +1,4 @@
-package ir.mseif.app.com.fzm;
+package ir.mseif.app.com.fzm.Services;
 
 import android.content.Context;
 import android.content.Intent;
@@ -6,17 +6,26 @@ import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.view.View;
 
+import ir.mseif.app.com.fzm.MainActivity;
+import ir.mseif.app.com.fzm.Activity.Map;
+import ir.mseif.app.com.fzm.R;
 import uk.co.chrisjenx.calligraphy.CalligraphyContextWrapper;
 
-public class Socket extends AppCompatActivity {
+public class Shutter_Door extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_socket);
+        setContentView(R.layout.activity_shutter__door);
     }
 
 
+
+    // Go To Time Activity
+    public void Go_To_Location (View v){
+        Intent intent = new Intent(getApplicationContext(), Map.class);
+        startActivity(intent);
+    }
 
 
     // Go To Time Activity
