+++ app/src/main/res/drawable/style_corner_white.xml
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <corners
+        android:radius="5dp" >
+    </corners>
+
+    <solid
+        android:color="@color/color_white" >
+    </solid>
+
+</shape>
