+++ app/src/main/res/layout/activity_asansor.xml
@@ -28,6 +28,7 @@
             android:background="@drawable/style_asansor_corner"
             android:drawableStart="@drawable/ic_location"
             android:text="آدرس دقیق"
+            android:onClick="Go_To_Location"
             android:textColor="@color/colorPrimary"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toTopOf="parent"
