+++ app/src/main/res/layout/activity_camera.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:focusable="true"
     android:focusableInTouchMode="true"
-    tools:context=".Camera">
+    tools:context=".Services.Camera">
 
     <android.support.constraint.ConstraintLayout
         android:layout_width="match_parent"
@@ -236,6 +236,7 @@
             android:paddingStart="15dp"
             android:paddingEnd="15dp"
             android:text="آدرس دقیق"
+            android:onClick="Go_To_Location"
             android:textColor="@color/colorPrimary"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
