+++ app/src/main/res/layout/activity_chandelier.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Chandelier">
+    tools:context=".Services.Chandelier">
 
     <android.support.constraint.ConstraintLayout
         android:layout_width="match_parent"
@@ -242,6 +242,7 @@
             android:paddingStart="15dp"
             android:paddingEnd="15dp"
             android:text="آدرس دقیق"
+            android:onClick="Go_To_Location"
             android:textColor="@color/colorPrimary"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
