+++ app/src/main/res/layout/activity_shutter__door.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Shutter_Door">
+    tools:context=".Services.Shutter_Door">
 
     <android.support.constraint.ConstraintLayout
         android:layout_width="match_parent"
@@ -15,12 +15,13 @@
         <Spinner
             android:id="@+id/spinner8"
             android:layout_width="match_parent"
-            android:layout_height="wrap_content"
+            android:layout_height="50dp"
             android:layout_marginStart="8dp"
             android:layout_marginLeft="8dp"
             android:layout_marginTop="8dp"
             android:layout_marginEnd="8dp"
             android:layout_marginRight="8dp"
+            android:background="@drawable/style_asansor_corner"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toTopOf="parent" />
@@ -28,12 +29,13 @@
         <Spinner
             android:id="@+id/spinner9"
             android:layout_width="match_parent"
-            android:layout_height="wrap_content"
+            android:layout_height="50dp"
             android:layout_marginStart="8dp"
             android:layout_marginLeft="8dp"
             android:layout_marginTop="8dp"
             android:layout_marginEnd="8dp"
             android:layout_marginRight="8dp"
+            android:background="@drawable/style_asansor_corner"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toBottomOf="@+id/spinner8" />
@@ -41,12 +43,13 @@
         <Spinner
             android:id="@+id/spinner10"
             android:layout_width="match_parent"
-            android:layout_height="wrap_content"
+            android:layout_height="50dp"
             android:layout_marginStart="8dp"
             android:layout_marginLeft="8dp"
             android:layout_marginTop="8dp"
             android:layout_marginEnd="8dp"
             android:layout_marginRight="8dp"
+            android:background="@drawable/style_asansor_corner"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toBottomOf="@+id/spinner9" />
@@ -210,6 +213,7 @@
             android:paddingStart="15dp"
             android:paddingEnd="15dp"
             android:text="آدرس دقیق"
+            android:onClick="Go_To_Location"
             android:textColor="@color/colorPrimary"
             tools:ignore="RtlCompat"
             app:layout_constraintEnd_toEndOf="parent"
